Added a simple memorize=TRUE option when initialize TileDB arrays.
This isn't quite correct as it doesn't preserve the tiled nature of the
original array, which means that it's not a guaranteed performance win.
Nonetheless, we implement it for consistency with the HDF5 case, and
besides, any cache inefficiency in the memorized array should be easily
exceeded by I/O bottlenecks in the original file-backed array.